Refactor: fetchJetpackSitePlugin() to be suspend function.
